Create stock.model
This model uses linear regression to predict the "Adj Close" column of the stock data, using the "Open", "Close", and "Volume" columns as features. The data is first split into training and testing sets, with 80% of the data being used for training and 20% for testing. The model is then trained on the training data and used to make predictions on the testing data. The mean squared error (MSE) is calculated as a measure of the model's performance, with a lower MSE indicating a better fit